// InvoicerClient/src/app/api/model/createEstimateCommand.ts
/**
 * Invoicer API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EstimateStatus } from './estimateStatus';
import { CreateEstimateProductItem } from './createEstimateProductItem';

export interface CreateEstimateCommand {
  companyId?: string;
  clientId?: string;
  estimateDate?: string;
  expiresOn?: string;
  status?: EstimateStatus;
  notes?: string | null;
  products?: Array<CreateEstimateProductItem> | null;
}
export namespace CreateEstimateCommand {}